Update README.md: add JupyterLab client
Thank you for maintaining this language server! Just wanted to let you know that we support CSS using your package, and add a shameless promotion to help users who may look for a client with CSS support. We take care of the configuration, so the only thing the user has to do is to install it via npm.